requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
langchain==0.0.335
langchain-openai==0.0.2
langchain-community==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
python-multipart==0.0.6
reportlab==4.0.7
aiofiles==23.2.1
httpx==0.25.2
Pillow==10.1.0